\name{chimera-package}
\alias{chimera-package}
\alias{chimera}
\docType{package}
\title{
A package for secondary analysis of fusion products
}
\description{
The package imports fusion results from tophat-fusion, mapSplice, deFuse, fusionmap, bellerophontes, fusionfinder, fusionhunter.
The package is made to facilitate the characterisation of fusion products events.
}
\details{
\tabular{ll}{
Package: \tab chimera\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-16\cr
License: \tab Artistic-2.0\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Raffaele A Calogero
Maintainer: Raffaele A Calogero <raffaele.calogero@unito.it>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
