\name{fSet}
\docType{class}
\alias{fSet-class}
\alias{fSet}
\alias{fusionData} 
\alias{fusionGRL} 
\alias{fusionRNA} 
\alias{addRNA}
\alias{fusionGA} 
\alias{addGA} 
\alias{fusionData,fSet-method}
\alias{fusionGRL,fSet-method}
\alias{fusionRNA,fSet-method}
\alias{addRNA,fSet-method}
\alias{fusionGA,fSet-method}
\alias{addGA,fSet-method} 
\alias{newfSet,fSet-method} 

\title{Class fSet, a class represent fusion data, and methods for
processing it}

\description{This is class representation for a fusion event.}

\section{Slots}{
   \describe{
      \item{fusionInfo}{A list:
		fusionTool: the tool that has generated the fusions
		UniqueCuttingPositionCount: the number of unique cutting positions detected for the fusion.
		SeedCount: the number of reads overlapping the break-point, i.e. spanning reads (FusionMap, FusionHunter, mapSplice, Tophat-fusion, ChimeraScan, STAR, Rsubread, FusionCatcher). Both spanning and encompassing reads (Bellerophontes, FusionFinder). Encompassing reads, i.e. one read of a pair on gene 1, and the other on gene2 (deFuse).
		RescuedCount: the number of reads overlapping the break-point rescued after identification of the break point (FusionMap). Encompassing reads (Tophat-fusionm Tophat-fusion-post, FusionCatcher, STAR). Both spanning and encompassing reads (ChimeraScan, Rsubread).
		SplicePattern: the splice pattern for a fusion junction
		FusionGene: the name of the fusion gene in the format gene1 -> gene2.
		frameShift: frameshift at break-point
	  }
      \item{fusionLoc}{A GRangesList encompassing fusion locations for gene 1 and 2}
      \item{fusionRNA}{A DNAStringSet encompassing the fusion transcript that can be generated by chimeraSeqs function.}
      \item{fusionGA}{A GAlignments object encompassing positions for all reads mapping on the DNAStringSet located in fusionRNA slot}
   }
}


\section{Methods}{
  Standard generic methods:
  \describe{
  \item{fusionData(fSet)}{An accessor function used to retrieve information for a fusion}
  \item{fusionGRL(fSet)}{An accessor function used to retrieve GRangesList encompassing fusion locations for gene 1 and 2}
  \item{fusionRNA(fSet)}{An accessor function used to extract the DNAStringSet}
  \item{addRNA(fSet, rna)}{An accessor function used to add the DNAStringSet to the fset object}
  \item{fusionGA(fSet)}{An accessor function used to extract the GAlignments object}
  \item{addGA(fSet, bam)}{An accessor function used to add the GAlignments object to the fSet object}
  }
 }
  
 \section{Constructor}{
 \describe{
     \item{newfSet}{
       \code{newfSet(fusionInfo = list(), 
	                 fusionLoc = GRangesList(),
                     fusionRNA = DNAStringSet(),
                     fusionGA = GAlignments())}:
       Creates a fSet object.
       \describe{
         \item{\code{fusionInfo}}{A list of the fusin characteristics see above slot fusionInfo}
         \item{\code{fusionLoc}}{A GRangesList encompassing the fusion break points}
         \item{\code{fusionRNA}}{A DNAStringSet encompassing the fusion transcript}
         \item{\code{fusionGA}}{A GAlignments encompassing the location of reads mapping on the fusion transcript}
       }
     }
   }
 }

\author{Raffaele A Calogero}
\examples{
#creating a fusion report from output of fusionMap
tmp <- importFusionData("fusionmap", paste(find.package(package="chimera"),"/examples/mcf7.FMFusionReport", sep=""), org="hg19")
#extracting the fusions names
fusion.names <- fusionName(tmp)
fusion.names
#extracting the fSet object ofr one of the fusions
myset <- tmp[[13]]
#constructing the fused sequence(s)
trs <- chimeraSeqs(myset, type="transcripts")
#adding the sequences to the fSet object
myset <- addRNA(myset , trs)
#extracting sequences from an fSet object
tmp.seq <- fusionRNA(myset)
#adding reads mapped on the fusion generated using tophatRun function
myset <- addGA(myset, paste(path.package(package="chimera"),"/examples/mcf7_trs_accepted_hits.bam",sep=""))
#extracting the GAlignments from an fSet object
ga <- fusionGA(myset)
}

\seealso{\code{\link{chimeraSeqs}}, \code{\link{oncofuseRun}}, \code{\link{fusionPeptides}}, \code{\link{prettyPrint}}}

\keyword{classes}

