\name{gapfillerInstallation}
\alias{gapfillerInstallation}
\title{A function to download a compiled version of GapFiller}
\description{A function allowing the download and installation of a compiled version of GapFiller in chimera package folder.
The source code of GapFiller can be retrieved at http://sourceforge.net/projects/gapfiller/files/?source=navbar .
The paper describing the tool is Nadalin F, Vezzi F, Policriti A. GapFiller: a de novo assembly approach to fill the gap within paired reads. BMC Bioinformatics. 2012;13 Suppl 14:S8.  
}
\usage{gapfillerInstallation(os=c("mac64","unix64"))}
\arguments{
  \item{os}{The supported operating systems}
}
\author{Raffaele A Calogero}
\examples{
#gapfillerInstallation(os="mac64")
}
\keyword{utilities}


