\name{MHmakeRandomString}
\alias{MHmakeRandomString}
\title{A function generating a random string}
\description{A function generating a random string.}
\usage{MHmakeRandomString()}
\value{
       a string
}
\author{Raffaele A Calogero}
\examples{
tmp.file <- paste(MHmakeRandomString(),".fa", sep="")
}
\keyword{utilities}


