\name{oncofuseInstallation}
\alias{oncofuseInstallation}
\title{A function to download oncofuse}
\description{A function allowing the download of oncofuse in the chimera folder. Oncofuse requires java}
\usage{oncofuseInstallation()}
\author{Raffaele A Calogero}
\examples{
#oncofuseInstallation()
}
\seealso{\code{\link{oncofuseRun}}}
\keyword{utilities}


