\name{picardInstallation}
\alias{picardInstallation}
\title{A function to download picard-tools}
\description{A function allowing the download of picard-tools in the chimera folder. Picard tools require java}
\usage{picardInstallation()}
\author{Raffaele A Calogero}
\examples{
#picardInstallation()
}
\keyword{utilities}


