\name{plotGO}
\alias{plotGO}
\title{A function relationship between different GO terms}
\description{A function relationship between different GO terms}
\usage{plotGO(go.term, go.class=c("BP","MF","CC"))}
\arguments{
  \item{go.term}{The GO term of interest, e.g. GO:0030154}
  \item{go.class}{GO classses}
}
\author{Raffaele A Calogero}
\examples{
#plotGO(go.term="GO:0030154",go.class="BP")
}
\keyword{utilities}


