\name{fSet}
\docType{class}
\alias{fSet-class}
\alias{fSet}
\alias{fusionData} 
\alias{fusionGRL} 
\alias{fusionRNA} 
\alias{addRNA}
\alias{fusionGA} 
\alias{addGA} 
\alias{fusionData,fSet-method}
\alias{fusionGRL,fSet-method}
\alias{fusionRNA,fSet-method}
\alias{addRNA,fSet-method}
\alias{fusionGA,fSet-method}
\alias{addGA,fSet-method} 

\title{Class fSet, a class represent fusion data, and methods for
processing it}

\description{This is class representation for a fusion event.}

\section{Slots}{
   \describe{
      \item{fusionInfo}{A list}
      \item{fusionLoc}{A GRangesList encompassing fusion locations for gene 1 and 2}
      \item{fusionRNA}{A DNAStringSet encompassing fusion transcripts that can be generated combining all transcripts for gene 1 and 2 the sequences are generated by the chimeraSeqs and they are needed to generate a bam file with tophatRun, which maps all the reads on the transcripts involved in the fusion. The bam file can be loaded in the slot fusionsLoc of the fSetSummary object as GappedAlignments. These data can be used to plot the reads coverage on the fused transcripts.}
      \item{fusionGA}{A GappedAlignments object encompassing positions for all reads mapping on the DNAStringSet located in fusionRNA slot}
   }
}


\section{Methods}{
  Standard generic methods:
  \describe{
  \item{fusionData (fSet)}{An accessor function used to retrieve information for a fusion}
  \item{fusionGRL (fSet)}{An accessor function used to retrieve GRangesList encompassing fusion locations for gene 1 and 2}
  \item{fusionRNA (fSet)}{An accessor function used to extract the DNAStringSet}
  \item{addRNA(fSet, rna)}{An accessor function used to add the DNAStringSet to the fset object}
  \item{fusionGA(fSet)}{An accessor function used to extract the GappedAlignments object}
  \item{addGA(fSet, bam)}{An accessor function used to add the GappedAlignments object to the fSet object}
  }
}
\author{Raffaele A Calogero}
\examples{
#creating a fusion report from output of fusionMap
tmp <- importFusionData("fusionmap", paste(find.package(package="chimera"),"/examples/mcf7.FMFusionReport", sep=""), org="hs")
#extracting the fusions names
fusion.names <- fusionName(tmp)
fusion.names
#extracting the fSet object ofr one of the fusions
myset <- tmp[[13]]
#constructing the fused sequence(s)
trs <- chimeraSeqs(myset, type="transcripts")
#adding the sequences to the fSet object
myset <- addRNA(myset , trs)
#extracting sequences from an fSet object
tmp.seq <- fusionRNA(myset)
#adding reads mapped on the fusion generated using tophatRun function
myset <- addGA(myset, paste(path.package(package="chimera"),"/examples/mcf7_trs_accepted_hits.bam",sep=""))
#extracting the GappedAlignments from an fSet object
ga <- fusionGA(myset)
}

\seealso{\code{\link{chimeraSeqs}}, \code{\link{tophatRun}}, \code{\link{plotCoverage}}}

\keyword{classes}
