\name{fusionPeptides}
\alias{fusionPeptides}
\title{A function associating to a fusion the peptides involved.}
\description{A function associate the donor and the acceptor peptides involved in the fusion.             
}
\usage{fusionPeptides(chimeraSeq.output, annotation="hsUCSC")}
\arguments{
  \item{chimeraSeq.output}{DNAStringSet encompassing the fusion event of interest, generated by chimeraSeq function}
  \item{annotation}{The annotation used to retrieve the UCSC names of the transcripts involved in the fusion}
}

\value{
An list encompassing:
\item{AAStringSet}{encompassing: fusion sequence, peptide from p1 and peptide from p2. In case the peptides are not in frame the AAStringSet will not contain the fusion sequence} 
\item{DNAStringSet}{encompassing the fusion transcript}
}
\author{Raffaele A Calogero}
\examples{
tmp <- importFusionData("fusionmap", paste(find.package(package="chimera"),"/examples/mcf7.FMFusionReport", sep=""), org="hs")
fusion.names <- fusionName(tmp)
fusion.names
myset <- tmp[1:3]
tmp.seq <- chimeraSeqSet(myset, parallel=FALSE)
tmpx <- lapply(tmp.seq,fusionPeptides)
}
\keyword{utilities}


