\name{gapfillerRun}
\alias{gapfillerRun}
\title{A function to confirm fusion break point by de novo assembly}
\description{A function that uses GapFiller to confirm, by de novo assembly, the presence of the fusion break point. The function needs as input the fusion transcript generated by chimeraSeqs function and two fastq files corresponding to the reads mapping over the fusion transcript.          
}
\usage{gapfillerRun(fusion.fa, seed1,  seed2, gapfiller=NULL, seed.ins=200, seed.var=50, 
block.length=5, overlap=20, max.length=5000, slack=30, 
k=6, global.mismatch=5)
}
\arguments{
  \item{fusion.fa}{fasta file with the fusin transcript}
  \item{seed1}{The R1 fastq of a pair-end}
  \item{seed2}{The R2 fastq of a pair-end}
  \item{gapfiller}{path to GapFiller executable program}
  \item{seed.ins}{seed reads insert size}
  \item{seed.var}{seed reads insert variation}
  \item{block.length}{length of perfect match}
  \item{overlap}{minimum suffix-prefix overlap}
  \item{max.length}{print only contigs <= max-length long}
  \item{slack}{number of overlaps: suffix-prefix overlap range is given by overlap, overlap + slack}
  \item{k}{length of the word used to hash}
  \item{global.mismatch}{maximum number of mismatches between mate and contig}
}
\value{
    The program will write in a temporary directory contigs.fasta and contig.stats, which are used to evaluate if the de novo  assembly allows the identification of the fusion break point. 
The function returns a list of three objects. The list is returned only in case that some of de novo assemblies cover the breakpoint junction. 
The list is made of: 
\item{contigs}{which is a PairwiseAlignments object} 
\item{junction.contigs}{which is a DNAStringSet encompassing the sequences present in the contigs object} 
\item{fusion}{which is a DNAStringSet object encompassing the fusion transcript}

}
\author{Raffaele A Calogero}
\examples{
#	tmp <- gapfillerRun(fusion.fa=paste(path.package("chimera", quiet = FALSE),"/examples/uc002xvp.1-243_uc002iyu.4-1031.fa",sep=""),seed1=paste(path.package("chimera", quiet = FALSE),"/examples/uc002xvp.1-243_uc002iyu.4-1031_R1.fastq",sep=""),  seed2=paste(path.package("chimera", quiet = FALSE),"/examples/uc002xvp.1-243_uc002iyu.4-1031_R1.fastq",sep=""))

}

\seealso{ \code{\link{chimeraSeqs}}, \code{\link{gapfillerInstallation}}}
\keyword{utilities}


