\name{importFusionData}
\alias{importFusionData}
\title{A function to import fusion data detected by different fusion finders}
\description{A function to import in a list fusions data detected by bellerophontes, defuse, fusionfinder, fusionhunter, mapsplice, tophat-fusion, fusionmap, chimerascan.             
}
\usage{importFusionData(format, filename, ...)}
\arguments{
  \item{format}{bellerophontes, defuse, fusionfinder, fusionhunter, mapsplice, tophat-fusion, fusionmap, chimerascan. Format allows to select the data structure to be imported}
  \item{filename}{The file generated by one of the fusion finders defined in format}
  \item{...}{Additional parameters}

}
\author{Raffaele A Calogero}
\examples{
    tmp <- importFusionData("fusionmap", paste(find.package(package="chimera"),"/examples/mcf7.FMFusionReport", sep=""), org="hs")
    #min.support allow to retrieve only the subset of fusions supported by a user defined minimal number of junction spanning reads
    #tmp <- importFusionData("chimerascan", "edgren_cs10.bedpe", min.support=10)
}
\keyword{utilities}

