\name{importFusionData}
\alias{importFusionData}
\title{A function to import fusion data detected by different fusion finders}
\description{A function to import in a list fusions data detected by bellerophontes, defuse, fusionfinder, fusionhunter, mapsplice, tophat-fusion, fusionmap, chimerascan.
In the case of chimerascan and star output it is possible to load the data applying a filter on the minimal number of reads supporting a specific fusion. Both chimerascan and star accept data generated using human hg19 and mouse mm9 reference. Star import function has a parallel option using BiocParallel package.             
}
\usage{importFusionData(format, filename, ...)}
\arguments{
  \item{format}{bellerophontes, defuse, fusionfinder, fusionhunter, mapsplice, tophat-fusion, fusionmap, chimerascan, star. Format allows to select the data structure to be imported}
  \item{filename}{The file generated by one of the fusion finders defined in format}
  \item{...}{Additional parameters}

}
\author{Raffaele A Calogero}
\examples{
    tmp <- importFusionData("fusionmap", paste(find.package(package="chimera"),"/examples/mcf7.FMFusionReport", sep=""), org="hs")
    #min.support allow to retrieve only the subset of fusions supported by a user defined minimal number of junction spanning reads
    #tmp <- importFusionData("chimerascan", "edgren_cs10.bedpe", min.support=10, org="hs")
    #download Edgren Chimeric.out.junction. This file encompass the results obtaines combined all cell lines used in Edgren experiment 
    #download.file("http://sourceforge.net/projects/ochguiextras/files/chimera/Chimeric.out.junction.zip/download", "Chimeric.out.junction.zip", mode="wb")
    #unzip("Chimeric.out.junction.zip")
    #tmp <- importFusionData("star", "Chimeric.out.junction", org="hs", parallel=T, hist.file=NULL, min.support=100)
}
\keyword{utilities}


